docs(rpc): batch requests page, examples and security rows for the batch id
