+++ container/website/app/components/content/go-generated/diagnostics-catalog.json
@@ -69,6 +69,56 @@
       "example": null
     },
     {
+      "code": "BAT001",
+      "subsystem": "markers",
+      "severity": "error",
+      "headline": "`batch()` element is not a route call the build can read ({0}); write `routes.a.b(...)` inline or bind it to a `const`/`let` in this file.",
+      "detail": "The build computes the batch id from the ORDERED list of route ids the\n`[...]` argument names, so every element must be a call on the client\nroutes proxy (`routes.users.getById(...)`) that it can trace statically,\neither written inline or bound to a `const` / `let` in the same file.\nA spread, a call on something that is not the routes proxy, a function\nparameter, or a binding without a route-call initializer cannot be read.\n\nFix: write the route call inline, or bind it first:\n-  batch([...prepared, routes.orders.list()]);\n+  const user = routes.users.getById(id);\n+  batch([user, routes.orders.list()]);",
+      "summary": null,
+      "fix": null,
+      "example": null
+    },
+    {
+      "code": "BAT002",
+      "subsystem": "markers",
+      "severity": "error",
+      "headline": "Batch `{0}` maps inputs differently in two places; the same ordered routes must always use the same `inputFrom()` mappings.",
+      "detail": "The batch id is derived from the ordered route ids alone, so every call\nsite naming that list shares one id, and the server executes one plan per\nid. Two sites with different `inputFrom()` sources, targets or mappers\nwould make the plan ambiguous.\n\nFix: make both sites use the same mappings, or give one of them a\ndifferent route order.",
+      "summary": null,
+      "fix": null,
+      "example": null
+    },
+    {
+      "code": "BAT003",
+      "subsystem": "markers",
+      "severity": "error",
+      "headline": "`inputFrom()` reads route `{0}` for route `{1}`, but the source is not in this batch or runs after the target; sources must be listed before the routes they feed.",
+      "detail": "Routes in a batch run in the order they are listed, and a route can only\nread the output of one that already ran. The source route must be an\nelement of the same `batch([...])` call and sit before the route whose\ninput it maps into.\n\nFix: list the source route first:\n-  batch([routes.orders.list(inputFrom(user, 'toUserId')), user]);\n+  batch([user, routes.orders.list(inputFrom(user, 'toUserId'))]);",
+      "summary": null,
+      "fix": null,
+      "example": null
+    },
+    {
+      "code": "BAT004",
+      "subsystem": "markers",
+      "severity": "error",
+      "headline": "Batch id `{0}` is shared by two different route lists; rename or reorder one of them so the ids no longer collide.",
+      "detail": "The batch id is a short hash of the ordered route ids. Two different\nlists hashing to the same id is very unlikely, but when it happens the\nserver could not tell the two plans apart.\n\nFix: change the route order of one batch, or split it into two batches.",
+      "summary": null,
+      "fix": null,
+      "example": null
+    },
+    {
+      "code": "BAT005",
+      "subsystem": "markers",
+      "severity": "error",
+      "headline": "`inputFrom()` mapper is not readable at build time ({0}); pass an inline arrow function or a string literal mapper name.",
+      "detail": "The build records which mapper feeds each batched route: an inline\nfunction is registered by content hash, a string names a mapper the\nserver registered. Anything else (a function reference, a computed\nstring, a value from a parameter) cannot be read statically.\n\nFix: inline the mapper or name it:\n-  inputFrom(user, pickId);\n+  inputFrom(user, (u) => u.id);\n+  inputFrom(user, 'toUserId');",
+      "summary": null,
+      "fix": null,
+      "example": null
+    },
+    {
       "code": "CTA001",
       "subsystem": "markers",
       "severity": "error",
