+++ container/website/content/01.rpc/01.introduction/01.about-mion-rpc.md
@@ -129,20 +129,20 @@ blur: 140px
     ---
     class: sm:col-span-2 lg:col-span-1 home-card
     ---
-    ### Routes Flows
-    [Execute multiple routes in a single HTTP request](/rpc/client/client-flow). Batch API calls together, and Orchestrate router logic from the client.
+    ### Batch Requests
+    [Execute multiple routes in a single HTTP request](/rpc/client/batch). Batch API calls together, and Orchestrate router logic from the client.
 
     <br>
 
     ::note{class="text-sm"}
-    [GraphQL-like data composition]{.text-highlighted} with the simplicity of RPC — resolve relationships in a single request using `serverMapFrom`.
+    [GraphQL-like data composition]{.text-highlighted} with the simplicity of RPC — resolve relationships in a single request using `inputFrom`.
     ::
     ::::
 
     ::::twoslash-code
     ---
-    path: packages/examples/src/_homepage/home-mapFrom.ts
-    title: mion-routesFlow.ts
+    path: packages/examples/src/_homepage/home-batch.ts
+    title: mion-batch.ts
     class: sm:col-span-2 lg:col-span-2
     ---
     ::::
