+++ container/website/content/01.rpc/02.server/09.security.md
@@ -12,6 +12,7 @@ Every request goes through the same steps, in this order, and each step refuses 
 
 | Step | What is checked | Error type |
 | --- | --- | --- |
+| Batch id | On a batch request, the id names a batch the build compiled into the server. Checked before the body is read | `batch-unknown-id` (404) |
 | Body size | The body (or the `?data=` query body) is not larger than `maxBodySize` | `request-payload-too-large` (413) |
 | Body shape | Valid JSON that is an object or an array, or a well-formed binary envelope | `parsing-json-request-error`, `invalid-request-body`, `binary-request-Deserialization-error` |
 | Query body | A `?data=` value is real base64url | `invalid-query-body` |
@@ -21,6 +22,10 @@ Every request goes through the same steps, in this order, and each step refuses 
 | Validation | Each parameter matches its declared type, and with `strictTypes` carries no unknown property | `validation-error` |
 
 Route ids, header names and body keys are looked up as own keys only, so a request naming `constructor` or `toString` is simply unknown. Keys named `__proto__`, `prototype` or `constructor` are refused everywhere a value is rebuilt from the wire.
+
+::tip
+A [batch](/rpc/client/batch) travels as its id alone. The routes it runs and the mappings between them never travel on the wire, so there is nothing for a caller to compose: the server refuses an unknown id before reading the body.
+::
 
 ::note
 Decoding runs before validation. That is why a decoder only converts values that already have the right wire shape, and leaves anything else for validation to refuse. The RunTypes guide explains the decoder contract in [Decoding Untrusted Input](/runtypes/guide/decoding-untrusted-input).
@@ -53,7 +58,7 @@ A string body is measured in characters, which is never more than its byte lengt
 
 ## Route Metadata Is Public
 
-The client needs to know how to encode a call and decode its answer, so the metadata route describes every route and every middleFn that takes parameters, takes headers or returns data. A raw middleFn, and a middleFn with neither parameters nor return data, never touch the wire and are not described. This is not an access control: route types are the public API contract, the same way a REST API's schema is.
+The client needs to know how to encode a call and decode its answer, so the metadata route describes every route and every middleFn that takes parameters, takes headers or returns data. A raw middleFn, and a middleFn with neither parameters nor return data, never touch the wire and are not described. It also lists the registered batch ids. This is not an access control: route types are the public API contract, the same way a REST API's schema is.
 
 ## What Stays Your Job
