+++ container/website/content/01.rpc/03.client/04.cancellation-timeouts.md
@@ -48,9 +48,9 @@ Use `client.abort()` to cancel every in-flight request at once, useful when a si
 `client.destroy()` also aborts all in-flight requests before clearing handlers.
 ::
 
-## Combined with Middleware Functions and RoutesFlow
+## Combined with Middleware Functions and Batches
 
-Cancellation and timeouts work the same with middleware functions and routes flows. Pass `signal` and `timeout` alongside the other setup options.
+Cancellation and timeouts work the same with middleware functions and [batches](/rpc/client/batch). Pass `signal` and `timeout` alongside the other setup options.
 
 ::code-split
 <code-import path="packages/examples/src/client/cancellation-with-middleFns.ts" lang="ts [client.ts]" />
@@ -58,7 +58,7 @@ Cancellation and timeouts work the same with middleware functions and routes flo
 ::
 
 ::tip
-In a routes flow the timeout applies to the whole batch request, not to each route inside it.
+In a batch the timeout applies to the whole request, not to each route inside it.
 ::
 
 ## Aborting Mutations
