+++ container/website/content/01.rpc/06.devtools/01.linter.md
@@ -13,7 +13,6 @@ mion ships an ESLint plugin with rules that catch common mistakes in mion routes
 | `strong-typed-routes` | Ensures handlers have explicit type annotations | ✅ Error |
 | `no-unreachable-union-types` | Detects unreachable union type members | ✅ Error |
 | `no-unsafe-property-names` | Refuses properties named `__proto__`, `prototype` or `constructor` | ✅ Error |
-| `no-vite-client` | Enforces constraints for client builds with no mion bundler preset | Opt-in |
 
 ## Installation
 
@@ -140,52 +139,6 @@ RunTypes purity checker and has been removed: `runtypes/pure-functions` from
 
 ---
 
-### `@mionjs/no-vite-client`
-
-The mion Vite and Next presets harvest inline `serverMapFrom` mappers from the client build. Any other build cannot, so there you must reference a server-registered pure function **by name** instead. This rule makes sure those name arguments are present and are string literals.
-
-This rule is **not** included in the recommended config — enable it only in projects that build the client without a mion preset.
-
-#### What it checks
-
-- **`serverMapFrom()`** must reference a server-registered pure function by a **string literal** name (its second argument) — inline mappers require the Vite or Next preset.
-- Name arguments must be **string literals**, not variables or expressions.
-
-#### Configuration
-
-```json [.eslintrc.json]
-{
-  "plugins": ["@mionjs"],
-  "rules": {
-    "@mionjs/no-vite-client": "error"
-  }
-}
-```
-
-#### ✅ Valid Example
-
-```ts
-import {serverMapFrom} from '@mionjs/client';
-
-// Reference a server-registered pure function by a string-literal name
-const userIdMapping = serverMapFrom(orderReq, 'orderToUserId');
-```
-
-#### ❌ Invalid Examples
-
-```ts
-import {serverMapFrom} from '@mionjs/client';
-
-// ❌ Inline mapper — can only be harvested by the Vite or Next preset
-serverMapFrom(orderReq, (order) => order.userId);
-
-// ❌ Name must be a string literal, not a variable
-const name = 'orderToUserId';
-serverMapFrom(orderReq, name);
-```
-
----
-
 ## RunTypes compiler rules (`runtypes/*`)
 
 mion's own rules above are **additive**: they sit alongside the 25 `runtypes/*` rules shipped by
@@ -260,7 +213,7 @@ silently drops or reshapes data — worth reading, not worth blocking on.
 ---
 
 ::note
-All rules except `no-vite-client` are included in the `plugin:@mionjs/recommended` configuration and are enabled by default with `error` severity. The `no-vite-client` rule is opt-in for projects that don't use Vite for client builds.
+All rules are included in the `plugin:@mionjs/recommended` configuration and are enabled by default with `error` severity.
 ::
 
 <!-- code-import-timestamp 1772057398380 -->
