+++ container/website/content/01.rpc/06.devtools/03.nextjs.md
@@ -29,15 +29,15 @@ result, nothing to change.
 
 Only your own TypeScript is rewritten. Files in `node_modules` are left alone.
 
-## Sharing mappers with your API
+## Sharing Batches With Your API
 
-If you use `serverMapFrom` to map data on the client, your Next build is the client half. It
-writes a manifest, and your mion API reads it.
+If your client sends [batches](/rpc/client/batch), your Next build is the client half. It gives
+every batch its id, writes the manifest, and your mion API reads it.
 
-<code-import path="packages/examples/src/codegen/next-config-server-mappers.ts" lang="ts [next.config.ts]" />
+<code-import path="packages/examples/src/codegen/next-config-batches.ts" lang="ts [next.config.ts]" />
 
-Point your API's Vite config at the same file with `serverMappers: {consume}` (see
-[Vite Configuration](/rpc/devtools/vite)) and the two halves line up.
+Point your API's Vite config at the same file with `batches: {consume}` (see
+[Batch Transport](/rpc/devtools/vite#batch-transport)) and the two halves line up.
 
 ## Editing types while the dev server runs
