+++ container/website/content/02.runtypes/02.guide/09.pure-functions.md
@@ -63,7 +63,7 @@ The named registrars want the `"namespace::name"` id written as a literal, which
 
 <code-import path="packages/examples/src/guide/anonymous-pure-fn.ts" lang="ts" />
 
-It mirrors the named lane's two forms. `registerAnonymousPureFn` takes the function directly, which is the shape a single-callback API like `serverMapFrom(t => t.id)` wants. When the helper needs one-time setup or wants to compose another pure fn, reach for `registerAnonymousPureFnFactory` and pass a factory instead.
+It mirrors the named lane's two forms. `registerAnonymousPureFn` takes the function directly, which is the shape a single-callback API like `inputFrom(order, o => o.id)` wants. When the helper needs one-time setup or wants to compose another pure fn, reach for `registerAnonymousPureFnFactory` and pass a factory instead.
 
 <code-import path="packages/examples/src/guide/anonymous-pure-fn-factory.ts" lang="ts" />
