+++ packages/examples/src/client/batch-orders.routes.ts
