+++ packages/examples/src/client/batches-server.vite.config.ts
@@ -4,10 +4,10 @@ import {mionVitePlugin} from '@mionjs/devtools/vite';
 
 export default defineConfig({
   plugins: [
-    // Consume (bake in) the mappers the client build harvested
+    // Consume (compile in) the batches the client build wrote
     mionVitePlugin({
-      serverMappers: {
-        consume: resolve(__dirname, '../client/.mion/server-mappers.json'),
+      batches: {
+        consume: resolve(__dirname, '../client/.mion/batches.json'),
       },
     }),
   ],
