+++ packages/examples/src/codegen/vite-client.config.ts
@@ -7,11 +7,11 @@ export default defineConfig({
   plugins: [
     mionVitePlugin({
       runTypes: {tsConfig: resolve(__dirname, 'tsconfig.json')},
-      // 1. serverMappers.emit — harvest the inline serverMapFrom mappers this client build
-      //    contains into a manifest the server build consumes. This is the core client<->server
-      //    transport (route validation/serialization metadata itself comes from the router
-      //    types at build time — no server needed for that).
-      serverMappers: {emit: resolve(__dirname, '.mion/server-mappers.json')},
+      // 1. batches.emit: write the batches this client build contains (their routes plus their
+      //    inline inputFrom mappers) into a manifest the server build consumes. This is the
+      //    client<->server batch transport (route validation/serialization metadata itself comes
+      //    from the router types at build time, no server needed for that).
+      batches: {emit: resolve(__dirname, '.mion/batches.json')},
       // 2. server (optional, e2e/dev only) — spawn the mion server as a child process and poll
       //    its port until it accepts connections (await `serverReady`), so client tests hit a
       //    live API. Only 'childProcess' mode exists.
